+++ build-logic/convention/src/main/kotlin/VKIDLibraryConventionPlugin.kt
@@ -2,7 +2,6 @@ import com.android.build.api.variant.LibraryAndroidComponentsExtension
 import com.android.build.gradle.LibraryExtension
 import com.vk.id.Versions
 import com.vk.id.configureAndroidLint
-import com.vk.id.configureDetekt
 import com.vk.id.configureKotest
 import com.vk.id.configureKotlinAndroid
 import com.vk.id.configureStrictMode
@@ -18,6 +17,7 @@ class VKIDLibraryConventionPlugin : Plugin<Project> {
             with(pluginManager) {
                 apply("com.android.library")
                 apply("org.jetbrains.kotlin.android")
+                apply("vkid.detekt")
             }
 
             extensions.configure<LibraryExtension> {
@@ -30,7 +30,6 @@ class VKIDLibraryConventionPlugin : Plugin<Project> {
             configureKotlinAndroid()
             configureStrictMode()
             configureKotest()
-            configureDetekt(isCompose = false)
             configureAndroidLint()
             dependencies {
                 add("androidTestImplementation", kotlin("test"))