+++ sdk/core/vkid/src/main/java/com/vk/id/internal/auth/AuthOptions.kt
@@ -24,7 +24,6 @@ internal data class AuthOptions(
 private const val APP_ID = "app_id"
 private const val CLIENT_ID = "client_id"
 
-// todo: Change to actual host after oauth2 completion
 private const val AUTHORITY_BROWSER = "id.vk.com"
 private const val AUTHORITY_CODE_FLOW = "vkcexternalauth-codeflow"
 private const val PATH_BROWSER = "authorize"