+++ sdk/core/vkid/src/main/java/com/vk/id/internal/auth/VKIDTokenPayload.kt
@@ -9,6 +9,7 @@ internal data class VKIDTokenPayload(
     val expiresIn: Long,
     val userId: Long,
     val state: String,
+    val scope: String,
 )
 
 /**